Improved tweeter
Passed in new args into `fortune`. The first one `-n 140` will
only return fortunes that are 140 chars in length. This remoed
the need to check the length of the `line` var. The seconde
being `-s` which only give fortuens that are "short" or 160
chars or less. I though that this might be a little redudnent
with the `-n 140` arg, except when fortuens with more than 200
chars were being passed into the `line` var.

Added logging function. Info about OAuth and HTTP status are
logged as well as the line that was tweeted and the lengthe
of the line.