Changed goe quote() method works
The quote() method is new recursive. The method will run
fortune, then get the length of the forturn and assign it
to a var. If that var is > 140 quote() is called agagin.
If that check return FALSE, the line is passed to a new
method, tweet(), along with the length of the line for
logging. This should hopefully take care of some issues
of the script not actually tweeting anyting.